divide app into components, manage state with simple contextAPI
