+++ src/components/Archive.jsx
@@ -1,0 +1,40 @@
+import React, { useContext, useState } from 'react'
+import { createRandomPost, PostContext } from '../context/PostContext';
+
+const Archive = () => {
+  const [posts] = useState(() =>
+    // 💥 WARNING: This might make your computer slow! Try a smaller `length` first
+    Array.from({ length: 10000 }, () => createRandomPost())
+  );
+
+  const [showArchive, setShowArchive] = useState(false);
+
+  const {onAddPost} = useContext(PostContext)
+
+  return (
+    <aside>
+      <h2>Post archive</h2>
+      <button onClick={() => setShowArchive((s) => !s)}>
+        {showArchive ? "Hide archive posts" : "Show archive posts"}
+      </button>
+
+      {showArchive && (
+        <ul>
+          {posts.map((post, i) => (
+            <li key={i}>
+              <p>
+                <strong>{post.title}:</strong> {post.body}
+              </p>
+              <button onClick={() => onAddPost(post)}>Add as new post</button>
+            </li>
+          ))}
+        </ul>
+      )}
+    </aside>
+  );
+}
+
+export default Archive
+
+
+
