+++ src/components/Footer.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const Footer = () => {
+  return <footer>&copy; by The Atomic Blog ✌️</footer>;
+}
+
+
+export default Footer
+
+
