+++ src/components/Header.jsx
@@ -1,0 +1,27 @@
+import React, { useContext } from 'react'
+import Results from './Results';
+import SearchPosts from './SearchPosts';
+import { PostContext } from '../context/PostContext';
+
+const Header = () => {
+  const {onClearPosts} = useContext(PostContext)
+  return (
+    <header>
+      <h1>
+        <span>⚛️</span>The Atomic Blog
+      </h1>
+      <div>
+        <Results />
+        <SearchPosts
+          
+         
+        />
+        <button onClick={onClearPosts}>Clear posts</button>
+      </div>
+    </header>
+  );
+}
+
+export default Header
+
+
