+++ src/components/List.jsx
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react'
+import { PostContext } from '../context/PostContext';
+
+const List = () => {
+  const {posts} = useContext(PostContext)
+  return (
+    <ul>
+      {posts.map((post, i) => (
+        <li key={i}>
+          <h3>{post.title}</h3>
+          <p>{post.body}</p>
+        </li>
+      ))}
+    </ul>
+  );
+}
+
+export default List
