+++ src/components/Main.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import Posts from './Posts';
+import FormAddPost from './FormAddPost';
+
+const Main = () => {
+  return (
+    <main>
+      <FormAddPost  />
+      <Posts />
+    </main>
+  );
+}
+
+export default Main
+
+
+
