+++ src/components/Posts.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import List from './List';
+
+const Posts = () => {
+  return (
+    <section>
+      <List />
+    </section>
+  );
+}
+
+export default Posts
+
+
+
+
+
