+++ src/components/SearchPosts.jsx
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react'
+import { PostContext } from '../context/PostContext';
+
+const SearchPosts = () => {
+  const {searchQuery, setSearchQuery} = useContext(PostContext)
+  return (
+    <input
+      value={searchQuery}
+      onChange={(e) => setSearchQuery(e.target.value)}
+      placeholder="Search posts..."
+    />
+  );
+}
+
+export default SearchPosts
+
+
+
